resolved 416 response that could cause exception to propogate out of storage layer
We were occasionally receiving 416 response messages when working
rapidly in a threaded environment on a single stream.  This actually
seems to be a race condition with the azure storage engine as a retry
with the same information will resolve the problem.  We have built the
retry into the storage engine to keep us from raising this issue when
there is no problem.  In the case where the problem persists we will
allow the exception to bubble out like it previously was after 100
attempts.  100 is a relatively aritrary number selected but would be
indicative of a corrupt message rather than an issue with azure blob
storage.