interum check-in working through a new model for maintaining commit history
Previously we were using the azure blob storage header to maintain
commit history and sequence number.  THe issue is the azure blob header
can only be 8K in size and this is for all headers combined.  we are
going to need to move our header logic into our primary repository.  The
current downside is a pretty significant potential performance loss for
detecting undispatched commits as we now need to interrogate each stream
individually.  We used to be able to make one large call to get all
metadata and make a decision from there.  We are now being forced to
come up with another scheme.

Right now we do not successfully execute, but we are close to being able
to rehydrate streams and be functional again, just at a poorer
performance rate.  This persistence mechanism is not backwards
compatable and prior aggregates will need to be destroyed (once it is
working).