working again
I have the storage engine working again, although I want to add some new
features to it.  We have paid a significant price for no longer having
the headers store our information.  Getting undispatched commits is now
growing timewise at a much faster rate.  I want to put a dirty flag in.
I also want to version the containers so if we change the engine again
we can up-convert or run the proper processor on it.  the core size of
the header has a large impact on performance also.  Working through
specifics on how to optimize still.