more optimizations
we reduced our retry count.  Optimized how we check if we are dirty
which has greatly increased our undispatched commit efficiency, although
there still seems to be a long way to go.  We still have a hard coded
header size which is giving us about 450 commits per aggregate max.  I
want this to be a flexible value or max at a more reasonable amount (for
instance 10,000 commits).  We will continue to iterate on this.  Also
still want to add versioning constructs to assist with future migrations
as we continue to optimize.