performance optimizations
we no longer fetchattributes as we get them in our original call to the
container now.  This allowed us to remove some major overhead and
improve performance via less web calls

fixed an issue where in getundispatchedcommits we were paying attention
to concurency issues when I don't think we shoudl care.  still digging
into this, but we should not let that fail us so easily