optimized our createifnotexistsandfetchattributes call
We used to make 2 API calls whenever we were checking if an aggregate
exists (even if it did exist).  We now only make one API call in the
case.  This was happening for every commit where we had to get a commit
sequence id and also for every commit.  so we have dropped our commit
calls by another 2 API calls.