additional optimizations and removing some configuration values
Removed some configuration values we no longer want to give users of the
API control over as it can have extremely detrimental impacts to the
system.  Users can no longer set the growth rate and the base size and
that can drastically increase the time to commit and get changes.

Removes a call to getheader that was no longer applicable.

Also simplified our getfrom call that gave a stream, min, and max
revision.  We no longer have two memorystreams that are managed and we
index directly into the byte[] array.  although more than likely not
noticable there is less logic which should give a minor bump in
performance.