minor optimization made to
We are in a state where we occasionally seem to corrupt an aggregate an
it no longer receives a correct guid id  we are also hitting some
concurrency issues when we have multiple streams writing in parrallel
(that should not happen at all).  This check in was to remove an
unnessesary write to a memory stream.  we could just advance the
position saving a little bit of write time.